adk_training/pz_bonus/copilot/course_generator/requirements.txt
# 🎓 GitHub Copilot Course Generator
# Dependencies
# Based on Google ADK 1.18.0

# ============================================================
# CORE ADK & GOOGLE AI
# ============================================================
google-adk==1.18.0
google-cloud-aiplatform[adk,agent_engines]==1.126.1
google-adk[eval]
google-genai
python-dotenv
toolbox-core

# ============================================================
# WEB SCRAPING & SEARCH
# ============================================================
beautifulsoup4>=4.12.0
requests>=2.31.0
aiohttp

# ============================================================
# FILE OPERATIONS & CONFIG
# ============================================================
pyyaml>=6.0.0
pydantic>=2.0.0

# ============================================================
# UTILITIES
# ============================================================
typing-extensions>=4.9.0
jinja2>=3.1.0
markdown>=3.5.0
colorlog>=6.8.0
tqdm>=4.66.0

